consolidate info into init.js
