+++ scripts/cronjob.js
@@ -9,6 +9,7 @@ const msgEmulator = require(path.join(__dirname, '..', 'src', 'msg-emulator'))
 // Create the real job from model job
 function createJob(job) {
   var time = new Date(job.pattern)
+  // when next phase complete, can just use date(job.pattern)
   var pattern = _.isNaN(time.getTime()) ? job.pattern : time
   new CronJob({
     cronTime: pattern,
@@ -21,6 +22,7 @@ Cronjob.all().then((jobs) => {
   _.each(jobs, createJob)
   // need to delete overdue
   // need to also create on add below
+  // also need to consider the best way to parse command
 })
 
 /* istanbul ignore next */
